# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Amar Pathagar API
  description: |
    Trust-Based Community Library System API
    
    A comprehensive API for managing a community-driven book sharing platform where trust and reputation matter.
  version: 1.0.0
  contact:
    name: NesoHQ
    url: https://github.com/nesohq
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.amarpathagar.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          type: string
          enum: [admin, member]
        avatar_url:
          type: string
        bio:
          type: string
        location_lat:
          type: number
          format: double
        location_lng:
          type: number
          format: double
        location_address:
          type: string
        success_score:
          type: integer
        books_shared:
          type: integer
        books_received:
          type: integer
        total_upvotes:
          type: integer
        total_downvotes:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Book:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        author:
          type: string
        isbn:
          type: string
        cover_url:
          type: string
        description:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        physical_code:
          type: string
        status:
          type: string
          enum: [available, reading, reserved, requested, on_hold]
        max_reading_days:
          type: integer
        current_holder_id:
          type: string
          format: uuid
        current_holder:
          $ref: '#/components/schemas/User'
        created_by:
          type: string
          format: uuid
        total_reads:
          type: integer
        average_rating:
          type: number
          format: float
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    HandoverThread:
      type: object
      properties:
        id:
          type: string
          format: uuid
        book_id:
          type: string
          format: uuid
        book:
          $ref: '#/components/schemas/Book'
        current_holder_id:
          type: string
          format: uuid
        current_holder:
          $ref: '#/components/schemas/User'
        next_holder_id:
          type: string
          format: uuid
        next_holder:
          $ref: '#/components/schemas/User'
        status:
          type: string
          enum: [active, completed, cancelled]
        handover_due_date:
          type: string
          format: date-time
        is_public:
          type: boolean
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    HandoverMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        thread_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/User'
        message:
          type: string
        is_system_message:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    ReadingHistory:
      type: object
      properties:
        id:
          type: string
          format: uuid
        book_id:
          type: string
          format: uuid
        book:
          $ref: '#/components/schemas/Book'
        reader_id:
          type: string
          format: uuid
        reader:
          $ref: '#/components/schemas/User'
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        duration_days:
          type: integer
        created_at:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
    Success:
      type: object
      properties:
        data:
          type: object
        message:
          type: string
paths:
...
  /users/{id}/profile:
    get:
      summary: Get user profile
      description: Get public profile of a user
      tags:
        - Users
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/profile:
    put:
      summary: Update user profile
      description: Update authenticated user's profile
      tags:
        - Users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                bio:
                  type: string
                avatar_url:
                  type: string
                location_address:
                  type: string
                location_lat:
                  type: number
                location_lng:
                  type: number
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/interests:
    post:
      summary: Add user interests
      description: Add reading interests for the user
      tags:
        - Users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - interests
              properties:
                interests:
                  type: array
                  items:
                    type: string
                  example: ["Science Fiction", "History", "Philosophy"]
      responses:
        '200':
          description: Interests added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
  /leaderboard:
    get:
      summary: Get leaderboard
      description: Get top users by success score
      tags:
        - Leaderboard
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Leaderboard data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
  /books:
    get:
      summary: List all books
      description: Get paginated list of books with optional filters
      tags:
        - Books
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [available, reading, reserved, requested, on_hold]
      responses:
        '200':
          description: List of books
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Book'
    post:
      summary: Create new book
      description: Add a new book to the library (admin only)
      tags:
        - Books
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - author
                - physical_code
              properties:
                title:
                  type: string
                author:
                  type: string
                isbn:
                  type: string
                description:
                  type: string
                category:
                  type: string
                physical_code:
                  type: string
                max_reading_days:
                  type: integer
                  default: 14
      responses:
        '201':
          description: Book created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Book'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /books/{id}/complete:
    post:
      summary: Mark book as completed
      description: Mark that you have finished reading the book
      tags:
        - Handover
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Book marked as completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /books/{id}/handover:
    get:
      summary: Get active handover thread
      description: Get the active handover thread for a book
      tags:
        - Handover
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Handover thread details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/HandoverThread'
        '404':
          description: No active handover thread found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /books/{id}/reading-status:
    get:
      summary: Get reading status
      description: Get extended reading history for current holder
      tags:
        - Handover
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Reading status
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ReadingHistory'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /handover/threads:
    get:
      summary: Get user's handover threads
      description: Get all handover threads for the current user
      tags:
        - Handover
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of handover threads
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/HandoverThread'
  /handover/threads/{id}/messages:
    get:
      summary: Get handover messages
      description: Get all messages in a handover thread
      tags:
        - Handover
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/HandoverMessage'

    post:
      summary: Post handover message
      description: Post a message to a handover thread
      tags:
        - Handover
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
      responses:
        '201':
          description: Message posted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...